Move SessionDuration and DefaultDuration to its own package for ease of categorization
